refactor mpremoteboard
