+++ src/mpflash/mpflash/list.py
@@ -6,12 +6,10 @@ from rich import print
 from rich.progress import track
 from rich.table import Table
 
-# TODO: - refactor so that we do not need the entire stubber package
-from .mpremoteboard.mpremoteboard import MPRemoteBoard
-
 from .cli_group import cli
 from .config import config
 from .logger import console, make_quiet
+from .mpremoteboard import MPRemoteBoard
 
 
 @cli.command("list", help="List the connected MCU boards.")
@@ -67,9 +65,9 @@ def show_mcus(
     """Show the list of connected boards in a nice table"""
     table = Table(
         title=title,
-        # expand=True,
         header_style="bold blue",
         collapse_padding=True,
+        width=100,
         # row_styles=["blue", "yellow"]
     )
     table.add_column("Serial", overflow="fold")
@@ -78,7 +76,7 @@ def show_mcus(
     table.add_column("Board", overflow="fold")
     # table.add_column("Variant") # TODO: add variant
     table.add_column("CPU")
-    table.add_column("Version", overflow="fold")
+    table.add_column("Version")
     table.add_column("build", justify="right")
 
     for mcu in track(conn_mcus, description="Updating board info", transient=True, update_period=0.1):
@@ -88,7 +86,7 @@ def show_mcus(
             except ConnectionError:
                 continue
         table.add_row(
-            mcu.serialport,
+            mcu.serialport.replace("/dev/", ""),
             mcu.family,
             mcu.port,
             mcu.board if mcu.board != "UNKNOWN" else mcu.description,
